+++ scripts/regenerate-protobuf-output.sh
@@ -8,7 +8,7 @@ root_dirpath="$(dirname "${script_dirpath}")"
 
 # ================================ CONSTANTS =======================================================
 GENERATOR_SCRIPT_FILENAME="generate-protobuf-bindings.sh"  # Must be on the PATH
-CORE_API_DIRNAME="core-api"
+CORE_API_DIRNAME="kurtosis-core-rpc-api"
 GOLANG_DIRNAME="golang"
 TYPESCRIPT_DIRNAME="typescript"
 
@@ -16,7 +16,7 @@ TYPESCRIPT_DIRNAME="typescript"
 input_dirpath="${root_dirpath}/${CORE_API_DIRNAME}"
 
 # Golang
-go_output_dirpath="${root_dirpath}/${GOLANG_DIRNAME}/core_api_bindings"
+go_output_dirpath="${root_dirpath}/${GOLANG_DIRNAME}/kurtosis_core_rpc_api_bindings"
 if ! GO_MOD_FILEPATH="${root_dirpath}/${GOLANG_DIRNAME}/go.mod" "${GENERATOR_SCRIPT_FILENAME}" "${input_dirpath}" "${go_output_dirpath}" golang; then
     echo "Error: An error occurred generating Go bindings in directory '${go_output_dirpath}'" >&2
     exit 1
@@ -24,7 +24,7 @@ fi
 echo "Successfully generated Go bindings in directory '${go_output_dirpath}'"
 
 # TypeScript
-typescript_output_dirpath="${root_dirpath}/${TYPESCRIPT_DIRNAME}/core_api_bindings"
+typescript_output_dirpath="${root_dirpath}/${TYPESCRIPT_DIRNAME}/kurtosis_core_rpc_api_bindings"
 if ! "${GENERATOR_SCRIPT_FILENAME}" "${input_dirpath}" "${typescript_output_dirpath}" typescript; then
     echo "Error: An error occurred generating TypeScript bindings in directory '${typescript_output_dirpath}'" >&2
     exit 1